Move ErrorLocation into syntax (break dependency)
