+++ berg-compiler/src/value/exception.rs
@@ -1,6 +1,6 @@
 use crate::syntax::{
-    AstIndex, AstRef, ByteRange, ExpressionRef, ExpressionTreeWalker, IdentifierIndex,
-    LineColumnRange,
+    ByteRange, ErrorLocation, ExpressionErrorPosition, ExpressionRef, ExpressionTreeWalker,
+    IdentifierIndex, LineColumnRange,
 };
 use crate::value::implement::*;
 use std::fmt;
@@ -31,49 +31,6 @@ pub struct Exception<'a> {
 #[derive(Debug, Clone)]
 pub struct CaughtException<'a>(Box<Exception<'a>>);
 
-#[derive(Debug, Clone)]
-pub enum ErrorLocation<'a> {
-    Generic,
-    SourceOnly(AstRef<'a>),
-    SourceExpression(AstRef<'a>, AstIndex),
-    SourceRange(AstRef<'a>, ByteRange),
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum ExpressionErrorPosition {
-    Expression,
-    Left,
-    Right,
-    LeftLeft,
-    LeftRight,
-    RightLeft,
-    RightRight,
-}
-
-impl ExpressionErrorPosition {
-    pub fn relative_to(self, new_position: ExpressionErrorPosition) -> ExpressionErrorPosition {
-        use ExpressionErrorPosition::*;
-        match (new_position, self) {
-            (new_position, Expression) => new_position,
-            (Expression, position) => position,
-            (Left, Left) => LeftLeft,
-            (Left, Right) => LeftRight,
-            (Right, Left) => RightLeft,
-            (Right, Right) => RightRight,
-            (LeftLeft, _)
-            | (LeftRight, _)
-            | (RightLeft, _)
-            | (RightRight, _)
-            | (_, LeftLeft)
-            | (_, LeftRight)
-            | (_, RightLeft)
-            | (_, RightRight) => unreachable!(
-                "Cannot reposition {:?} on top of {:?}: too deep!",
-                self, new_position
-            ),
-        }
-    }
-}
 impl<'a> EvalException<'a> {
     pub fn reposition(self, new_position: ExpressionErrorPosition) -> EvalException<'a> {
         use EvalException::*;
